Address PR review: fix extra query in IsGameManager, move fixtures to conftest, move transfer logic to GameManager
- Add select_related("managing_member") to resolve_game() so IsGameManager.has_permission does not issue an extra query when managing_member_id is set
- Move transfer_management_if_needed logic to GameManager.transfer_management_if_needed(game, exiting_user, reason); keep a thin delegation method on Game for instance callers (phase/models.py, member/views.py)
- Update all test call sites to use Game.objects.transfer_management_if_needed(game, ...) per the manager pattern
- Move player_run_active_game_factory, player_run_pending_game_factory, game_master_game_factory, and management_transfer_sandbox_game_factory fixtures from test_management_transfer.py to game/tests/conftest.py; rename sandbox fixture to avoid collision with root conftest sandbox_game_factory
